# Display optimizations (between 2x00 and 8x00 times faster) (ignore, superseed by #160)