--- a/full_benchmark.sh
+++ b/full_benchmark.sh
@@ -1,9 +1,21 @@
 #!/usr/bin/bash
 
-if [ "$#" != "1" ]; then
+if [ "$#" != "1" -a "$#" != "2" ]; then
     echo "Usage benchmark.sh <outputfile>"
+    echo "Usage benchmark.sh <outputfile> tracemalloc"
     exit 1
 fi
+
+if [ "$2" = "tracemalloc" ]; then
+    tracemalloc="--tracemalloc"
+elif [ "$2" = "" ]; then
+    tracemalloc=""
+else
+    echo "Usage benchmark.sh <outputfile>"
+    echo "Usage benchmark.sh <outputfile> tracemalloc"
+    exit 1
+fi
+
 outputfile=$1
 
 if [ ! -f benchmark.py ]; then
@@ -15,25 +27,25 @@ for inputfile in $(ls -1 tests/captured/*.input); do
     export GEOMETRY=24x80
     echo "$inputfile - $GEOMETRY"
     echo "======================"
-    BENCHMARK=$inputfile python benchmark.py --append $outputfile
+    BENCHMARK=$inputfile python benchmark.py $tracemalloc --append $outputfile
 
     export GEOMETRY=240x800
     echo "$inputfile - $GEOMETRY"
     echo "======================"
-    BENCHMARK=$inputfile python benchmark.py --append $outputfile
+    BENCHMARK=$inputfile python benchmark.py $tracemalloc --append $outputfile
 
     export GEOMETRY=2400x8000
     echo "$inputfile - $GEOMETRY"
     echo "======================"
-    BENCHMARK=$inputfile python benchmark.py --append $outputfile
+    BENCHMARK=$inputfile python benchmark.py $tracemalloc --append $outputfile
 
     export GEOMETRY=24x8000
     echo "$inputfile - $GEOMETRY"
     echo "======================"
-    BENCHMARK=$inputfile python benchmark.py --append $outputfile
+    BENCHMARK=$inputfile python benchmark.py $tracemalloc --append $outputfile
 
     export GEOMETRY=2400x80
     echo "$inputfile - $GEOMETRY"
     echo "======================"
-    BENCHMARK=$inputfile python benchmark.py --append $outputfile
+    BENCHMARK=$inputfile python benchmark.py $tracemalloc --append $outputfile
 done